+++ .env.example
@@ -3,26 +3,32 @@ GOERLI_DEPLOYER_PRIVATE_KEY=
 LOCAL_DEPLOYER_PRIVATE_KEY=
 
 # rpc urls
-API_KEY_NODE_REAL=
 MAINNET_RPC_URL=
 # Fork tests can run against a mainnet node or a local fork (anvil)
 FORK_RPC_URL=$MAINNET_RPC_URL
 # FORK_RPC_URL=http://localhost:8545
-PROVIDER_URL_TESTNET=
-POLYGON_ZKEVM_RPC_URL=https://zkevm-rpc.com
-
-# chain explorer api key for contract verification
-GOERLI_ETHERSCAN_API_KEY=
-ETHERSCAN_API_KEY=
-POLYSCAN_ZKEVM_API_KEY=
-# used by hardhat
-MAINNET_ETHERSCAN_API_KEY=
-
-MAX_DEPOSIT_AMOUNT=0
-
-# Open Zeppelin Defender keys
-DEFENDER_KEY=
-DEFENDER_SECRET=
 
 # Fix the block to fork from when running hardhat forks
 BLOCK_NUMBER=
+
+# chain explorer api key for contract verification
+ETHERSCAN_API_KEY=
+GOERLI_ETHERSCAN_API_KEY=
+
+# Open Zeppelin Defender Team API key
+DEFENDER_TEAM_KEY=
+DEFENDER_TEAM_SECRET=
+
+# Open Zeppelin Defender Relayer account API key
+DEFENDER_RELAYER_KEY=
+DEFENDER_RELAYER_SECRET=
+
+# Used by the depositAllEL Hardhat task
+MAX_DEPOSIT_AMOUNT=0
+
+# Used to impersonate accounts when running Hardhat tasks
+IMPERSONATE=
+# IMPERSONATE=0x5De069482Ac1DB318082477B7B87D59dfB313f91    # Relayer account
+# IMPERSONATE=0xEc574b7faCEE6932014EbfB1508538f6015DCBb0    # MultiSig
+# IMPERSONATE=0xF83cacA1bC89e4C7f93bd17c193cD98fEcc6d758    # ProxyAdmin
+# IMPERSONATE=0xc055De577ce2039E6D35621E3a885df9Bb304AB9    # primeETH whale
