+++ script/utils/logger.js
@@ -3,10 +3,10 @@ const debug = require("debug");
 /**
  * Creates a logger for a module.
  * @example
- *   const log = require("../utils/logger")("task:OSwap");
+ *   const log = require("../utils/logger")("task:deposits");
  *   log('something interesting happened');
- * @param {string} module name of the module to log for. eg "task:OSwap:snap"
+ * @param {string} module name of the module to log for. eg "task:deposits"
  */
-const logger = (module) => debug(`origin:${module}`);
+const logger = (module) => debug(`prime:${module}`);
 
 module.exports = logger;
