+++ script/utils/signers.js
@@ -35,18 +35,14 @@ async function getSigner(address = undefined) {
   if (process.env.IMPERSONATE) {
     let address = process.env.IMPERSONATE;
     if (!address.match(ethereumAddress)) {
-      throw Error(
-        `Environment variable IMPERSONATE is an invalid Ethereum address or contract name`
-      );
+      throw Error(`Environment variable IMPERSONATE is an invalid Ethereum address or contract name`);
     }
-    log(
-      `Impersonating account ${address} from IMPERSONATE environment variable`
-    );
+    log(`Impersonating account ${address} from IMPERSONATE environment variable`);
     return await impersonateAndFund(address);
   }
 
   // If using Defender Relayer
-  if (process.env.DEFENDER_API_KEY && process.env.DEFENDER_API_SECRET) {
+  if (process.env.DEFENDER_RELAYER_KEY && process.env.DEFENDER_RELAYER_SECRET) {
     return await getDefenderSigner();
   }
 
@@ -67,20 +63,20 @@ const getDefenderSigner = async () => {
   const speed = process.env.SPEED || "fastest";
   if (!["safeLow", "average", "fast", "fastest"].includes(speed)) {
     console.error(
-      `Defender Relay Speed param must be either 'safeLow', 'average', 'fast' or 'fastest'. Not "${speed}"`
+      `Defender Relay Speed param must be either 'safeLow', 'average', 'fast' or 'fastest'. Not "${speed}"`,
     );
     process.exit(2);
   }
   const credentials = {
-    relayerApiKey: process.env.DEFENDER_API_KEY,
-    relayerApiSecret: process.env.DEFENDER_API_SECRET,
+    relayerApiKey: process.env.DEFENDER_RELAYER_KEY,
+    relayerApiSecret: process.env.DEFENDER_RELAYER_SECRET,
   };
   const client = new Defender(credentials);
   const provider = client.relaySigner.getProvider();
 
   const signer = client.relaySigner.getSigner(provider, { speed });
   log(
-    `Using Defender Relayer account ${await signer.getAddress()} from env vars DEFENDER_API_KEY and DEFENDER_API_SECRET`
+    `Using Defender Relayer account ${await signer.getAddress()} from env vars DEFENDER_RELAYER_KEY and DEFENDER_RELAYER_SECRET`,
   );
   return signer;
 };
